Fix various urls/paths used throughout templates to work with the new routes, including getting rid of some hardcoded urls.
There's one slight inconsistency with the new routes, which is the path of the
solution creation route. For all the other resource-y things, the 'new' url is
always the same as the 'create' url, but because the solution urls were
generated automatically (using resources) they're a bit limited and have ended
up being different (the url for 'new' as 'new' at the end of it, whereas the
url for 'create' doesn't and is simply the same url as the url for 'index').