Added simple tests
I still can't find a way of running ruby in the terminal, which this
requires. Going to 52.16.192.137:3000 does get me to the website, but
the tests cannot be run from there. Can someone run them for me with $
bin/rake test test/models/user_test.rb and see if they work